move graph rotation optimizer to dedicated file
